+++ packages/ui-positions/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @orderly.network/ui-positions
 
+## 2.8.14
+
+### Patch Changes
+
+- publish
+- 6355a39: publish
+- Updated dependencies
+- Updated dependencies [6355a39]
+  - @orderly.network/react-app@2.8.14
+  - @orderly.network/hooks@2.8.14
+  - @orderly.network/i18n@2.8.14
+  - @orderly.network/perp@4.8.14
+  - @orderly.network/types@2.8.14
+  - @orderly.network/ui@2.8.14
+  - @orderly.network/ui-connector@2.8.14
+  - @orderly.network/ui-leverage@2.8.14
+  - @orderly.network/ui-share@2.8.14
+  - @orderly.network/ui-tpsl@2.8.14
+  - @orderly.network/utils@2.8.14
+
+## 2.8.14-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react-app@2.8.14-alpha.0
+  - @orderly.network/hooks@2.8.14-alpha.0
+  - @orderly.network/i18n@2.8.14-alpha.0
+  - @orderly.network/perp@4.8.14-alpha.0
+  - @orderly.network/types@2.8.14-alpha.0
+  - @orderly.network/ui@2.8.14-alpha.0
+  - @orderly.network/ui-connector@2.8.14-alpha.0
+  - @orderly.network/ui-leverage@2.8.14-alpha.0
+  - @orderly.network/ui-share@2.8.14-alpha.0
+  - @orderly.network/ui-tpsl@2.8.14-alpha.0
+  - @orderly.network/utils@2.8.14-alpha.0
+
 ## 2.8.13
 
 ### Patch Changes